Refactor logical design
